+++ lib/netlink/attr_nst.c
@@ -1,4 +1,4 @@
-#include <bits/null.h>
+#include <cdefs.h>
 
 #include "base.h"
 #include "attr.h"
@@ -8,7 +8,7 @@ static int extend_to_4bytes(int n)
 	return n + ((4 - (n & 3)) & 3);
 }
 
-int nl_check_nest(char* buf, int len)
+int nl_check_nest(char* buf, size_t len)
 {
 	char* end = buf + len;
